Extract validator operator inventory helpers
